third_party/blink/web_tests/external/wpt/editing/run/inserthtml-expected.txt

This is a testharness.js-based test.
Found 58 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["stylewithcss","true"],["inserthtml","ab<b>c</b>d"]] "foo[]bar" queryCommandState("stylewithcss") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["defaultparagraphseparator","div"],["inserthtml","<p>abc"]] "<p>foo[bar]baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p>abc</p><p>baz</p>" but got "<p>fooabcbaz</p>"
[FAIL] [["defaultparagraphseparator","p"],["inserthtml","<p>abc"]] "<p>foo[bar]baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p>abc</p><p>baz</p>" but got "<p>fooabcbaz</p>"
[FAIL] [["defaultparagraphseparator","div"],["inserthtml","<li>abc"]] "<p>foo[bar]baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><div>abc</div><p>baz</p>" but got "<p>foo</p><li>abc</li><p>baz</p>"
[FAIL] [["defaultparagraphseparator","p"],["inserthtml","<li>abc"]] "<p>foo[bar]baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p>abc</p><p>baz</p>" but got "<p>foo</p><li>abc</li><p>baz</p>"
[FAIL] [["defaultparagraphseparator","div"],["inserthtml","<p>abc"]] "<ol>{<li>foo</li>}<li>bar</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>abc</p><ol><li>bar</li></ol>" but got "<ol><li>abcbar</li></ol>"
[FAIL] [["defaultparagraphseparator","p"],["inserthtml","<p>abc"]] "<ol>{<li>foo</li>}<li>bar</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>abc</p><ol><li>bar</li></ol>" but got "<ol><li>abcbar</li></ol>"
[FAIL] [["defaultparagraphseparator","div"],["inserthtml","<p>abc"]] "<ol><li>foo</li>{<li>bar</li>}<li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li></ol><p>abc</p><ol><li>baz</li></ol>" but got "<ol><li>foo</li><li>abcbaz</li></ol>"
[FAIL] [["defaultparagraphseparator","p"],["inserthtml","<p>abc"]] "<ol><li>foo</li>{<li>bar</li>}<li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li></ol><p>abc</p><ol><li>baz</li></ol>" but got "<ol><li>foo</li><li>abcbaz</li></ol>"
[FAIL] [["inserthtml","abc"]] "<script>f[o]o</script>bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<script>fabco</script>bar" but got "<script>foo</script>abcbar"
[FAIL] [["stylewithcss","true"],["inserthtml","<b>abc</b>"]] "<script>f[o]o</script>bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<script>f<b>abc</b>o</script>bar" but got "<script>foo</script><b>abc</b>bar"
[FAIL] [["stylewithcss","false"],["inserthtml","<b>abc</b>"]] "<script>f[o]o</script>bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<script>f<b>abc</b>o</script>bar" but got "<script>foo</script><b>abc</b>bar"
[FAIL] [["inserthtml","<a>abc</a>"]] "<a>f[o]o</a>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<a>f</a><a>abc</a><a>o</a>" but got "<a>f<a>abc</a>o</a>"
[FAIL] [["inserthtml","<a href=/>abc</a>"]] "<a href=.>f[o]o</a>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<a href=\\".\\">f</a><a href=\\"/\\">abc</a><a href=\\".\\">o</a>" but got "<a href=\\".\\">f<a href=\\"/\\">abc</a>o</a>"
[FAIL] [["inserthtml","<dt>abc</dt>"]] "<dl><dt>f[o]o<dd>bar</dl>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<dl><dt>f</dt><dt>abc</dt><dt>o</dt><dd>bar</dd></dl>" but got "<dl><dt>f<dt>abc</dt>o</dt><dd>bar</dd></dl>"
[FAIL] [["inserthtml","<dt>abc</dt>"]] "<dl><dt>foo<dd>b[a]r</dl>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<dl><dt>foo</dt><dd>b</dd><dt>abc</dt><dd>r</dd></dl>" but got "<dl><dt>foo</dt><dd>b<dt>abc</dt>r</dd></dl>"
[FAIL] [["inserthtml","<dd>abc</dd>"]] "<dl><dt>f[o]o<dd>bar</dl>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<dl><dt>f</dt><dd>abc</dd><dt>o</dt><dd>bar</dd></dl>" but got "<dl><dt>f<dd>abc</dd>o</dt><dd>bar</dd></dl>"
[FAIL] [["inserthtml","<dd>abc</dd>"]] "<dl><dt>foo<dd>b[a]r</dl>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<dl><dt>foo</dt><dd>b</dd><dd>abc</dd><dd>r</dd></dl>" but got "<dl><dt>foo</dt><dd>b<dd>abc</dd>r</dd></dl>"
[FAIL] [["inserthtml","<dt>abc</dt>"]] "f[o]o" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "f<dl><dt>abc</dt></dl>o" but got "f<dt>abc</dt>o"
[FAIL] [["inserthtml","<dt>abc</dt>"]] "<ol><li>f[o]o</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>f<dl><dt>abc</dt></dl>o</li></ol>" but got "<ol><li>f<dt>abc</dt>o</li></ol>"
[FAIL] [["inserthtml","<dd>abc</dd>"]] "f[o]o" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "f<dl><dd>abc</dd></dl>o" but got "f<dd>abc</dd>o"
[FAIL] [["inserthtml","<dd>abc</dd>"]] "<ol><li>f[o]o</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>f<dl><dd>abc</dd></dl>o</li></ol>" but got "<ol><li>f<dd>abc</dd>o</li></ol>"
[FAIL] [["inserthtml","<li>abc</li>"]] "<dir><li>f[o]o</dir>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<dir><li>f</li><li>abc</li><li>o</li></dir>" but got "<dir><li>f<li>abc</li>o</li></dir>"
[FAIL] [["inserthtml","<li>abc</li>"]] "<ol><li>f[o]o</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>f</li><li>abc</li><li>o</li></ol>" but got "<ol><li>f<li>abc</li>o</li></ol>"
[FAIL] [["inserthtml","<li>abc</li>"]] "<ul><li>f[o]o</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>f</li><li>abc</li><li>o</li></ul>" but got "<ul><li>f<li>abc</li>o</li></ul>"
[FAIL] [["inserthtml","<ol><li>abc</ol>"]] "<dir><li>f[o]o</dir>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<dir><li>f<ol><li>abc</li></ol>o</li></dir>" but got "<dir><li>f</li><li>abc</li><li>o</li></dir>"
[FAIL] [["inserthtml","<ol><li>abc</ol>"]] "<ol><li>f[o]o</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>f<ol><li>abc</li></ol>o</li></ol>" but got "<ol><li>f</li><li>abc</li><li>o</li></ol>"
[FAIL] [["inserthtml","<ol><li>abc</ol>"]] "<ul><li>f[o]o</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>f<ol><li>abc</li></ol>o</li></ul>" but got "<ul><li>f</li><li>abc</li><li>o</li></ul>"
[FAIL] [["inserthtml","<ul><li>abc</ul>"]] "<dir><li>f[o]o</dir>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<dir><li>f<ul><li>abc</li></ul>o</li></dir>" but got "<dir><li>f</li><li>abc</li><li>o</li></dir>"
[FAIL] [["inserthtml","<ul><li>abc</ul>"]] "<ol><li>f[o]o</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>f<ul><li>abc</li></ul>o</li></ol>" but got "<ol><li>f</li><li>abc</li><li>o</li></ol>"
[FAIL] [["inserthtml","<ul><li>abc</ul>"]] "<ul><li>f[o]o</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>f<ul><li>abc</li></ul>o</li></ul>" but got "<ul><li>f</li><li>abc</li><li>o</li></ul>"
[FAIL] [["defaultparagraphseparator","div"],["inserthtml","<li>abc</li>"]] "f[o]o" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "f<div>abc</div>o" but got "f<li>abc</li>o"
[FAIL] [["defaultparagraphseparator","p"],["inserthtml","<li>abc</li>"]] "f[o]o" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "f<p>abc</p>o" but got "f<li>abc</li>o"
[FAIL] [["inserthtml","<nobr>abc</nobr>"]] "<nobr>f[o]o</nobr>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<nobr>f</nobr><nobr>abc</nobr><nobr>o</nobr>" but got "<nobr>f<nobr>abc</nobr>o</nobr>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["inserthtml","<p>abc"]] "<font color=blue>foo[]bar</font>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<font color=\\"blue\\">foo</font><p><span style=\\"color:rgb(0, 0, 255)\\">abc</span></p><font color=\\"blue\\">bar</font>" but got "<font color=\\"blue\\">foo</font>abc<span style=\\"color:rgb(0, 0, 255)\\">bar</span><br>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["inserthtml","<p>abc"]] "<font color=blue>foo[]bar</font>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<font color=\\"blue\\">foo</font><p><font color=\\"#0000ff\\">abc</font></p><font color=\\"blue\\">bar</font>" but got "<font color=\\"blue\\">foo</font>abc<span style=\\"color:rgb(0, 0, 255)\\">bar</span><br>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["inserthtml","<p>abc"]] "<font color=blue>foo[]bar</font>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<font color=\\"blue\\">foo</font><p><span style=\\"color:rgb(0, 0, 255)\\">abc</span></p><font color=\\"blue\\">bar</font>" but got "<font color=\\"blue\\">foo</font>abc<span style=\\"color:rgb(0, 0, 255)\\">bar</span><br>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["inserthtml","<p>abc"]] "<font color=blue>foo[]bar</font>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<font color=\\"blue\\">foo</font><p><font color=\\"#0000ff\\">abc</font></p><font color=\\"blue\\">bar</font>" but got "<font color=\\"blue\\">foo</font>abc<span style=\\"color:rgb(0, 0, 255)\\">bar</span><br>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["inserthtml","<p>abc"]] "<span style=color:blue>foo[]bar</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"color:rgb(0, 0, 255)\\">foo</span><p><span style=\\"color:rgb(0, 0, 255)\\">abc</span></p><span style=\\"color:rgb(0, 0, 255)\\">bar</span>" but got "<span style=\\"color:rgb(0, 0, 255)\\">foo</span>abc<span style=\\"color:rgb(0, 0, 255)\\">bar</span><br>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["inserthtml","<p>abc"]] "<span style=color:blue>foo[]bar</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"color:rgb(0, 0, 255)\\">foo</span><p><font color=\\"#0000ff\\">abc</font></p><span style=\\"color:rgb(0, 0, 255)\\">bar</span>" but got "<span style=\\"color:rgb(0, 0, 255)\\">foo</span>abc<span style=\\"color:rgb(0, 0, 255)\\">bar</span><br>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["inserthtml","<p>abc"]] "<span style=color:blue>foo[]bar</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"color:rgb(0, 0, 255)\\">foo</span><p><span style=\\"color:rgb(0, 0, 255)\\">abc</span></p><span style=\\"color:rgb(0, 0, 255)\\">bar</span>" but got "<span style=\\"color:rgb(0, 0, 255)\\">foo</span>abc<span style=\\"color:rgb(0, 0, 255)\\">bar</span><br>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["inserthtml","<p>abc"]] "<span style=color:blue>foo[]bar</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"color:rgb(0, 0, 255)\\">foo</span><p><font color=\\"#0000ff\\">abc</font></p><span style=\\"color:rgb(0, 0, 255)\\">bar</span>" but got "<span style=\\"color:rgb(0, 0, 255)\\">foo</span>abc<span style=\\"color:rgb(0, 0, 255)\\">bar</span><br>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["inserthtml","<p>abc"]] "<span style=font-variant:small-caps>foo[]bar</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"font-variant:small-caps\\">foo</span><p>abc</p><span style=\\"font-variant:small-caps\\">bar</span>" but got "<span style=\\"font-variant:small-caps\\">fooabcbar<br></span>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["inserthtml","<p>abc"]] "<span style=font-variant:small-caps>foo[]bar</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"font-variant:small-caps\\">foo</span><p>abc</p><span style=\\"font-variant:small-caps\\">bar</span>" but got "<span style=\\"font-variant:small-caps\\">fooabcbar<br></span>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["inserthtml","<p>abc"]] "<span style=font-variant:small-caps>foo[]bar</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"font-variant:small-caps\\">foo</span><p>abc</p><span style=\\"font-variant:small-caps\\">bar</span>" but got "<span style=\\"font-variant:small-caps\\">fooabcbar<br></span>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["inserthtml","<p>abc"]] "<span style=font-variant:small-caps>foo[]bar</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"font-variant:small-caps\\">foo</span><p>abc</p><span style=\\"font-variant:small-caps\\">bar</span>" but got "<span style=\\"font-variant:small-caps\\">fooabcbar<br></span>"
[FAIL] [["inserthtml"," "]] "<p>[foo]</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p> <br></p>" but got "<p> </p>"
[FAIL] [["inserthtml","abc"]] "<p><!--foo-->{}<span><br></span><!--bar--></p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><!--foo-->abc<!--bar--></p>" but got "<p><!--foo-->abc<span></span><!--bar--></p>"
[FAIL] [["inserthtml","<!--abc-->"]] "<p><!--foo-->{}<span><br></span><!--bar--></p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><!--foo--><!--abc--><!--bar--><br></p>" but got "<p><!--foo--><!--abc--><span></span><!--bar--></p>"
[FAIL] [["inserthtml","abc"]] "<p>{}<span><!--foo--><br><!--bar--></span></p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>abc</p>" but got "<p>abc<span><!--foo--><!--bar--></span></p>"
[FAIL] [["inserthtml","<!--abc-->"]] "<p>{}<span><!--foo--><br><!--bar--></span></p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><!--abc--><br></p>" but got "<p><!--abc--><span><!--foo--><!--bar--></span></p>"
[FAIL] [["inserthtml","<!--abc-->"]] "<p><!--foo--><span><br></span>{}<!--bar--></p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><!--foo--><span><!--abc--><br></span><!--bar--></p>" but got "<p><!--foo--><!--abc--><span></span><!--bar--></p>"
[FAIL] [["inserthtml","abc"]] "<p><span><!--foo--><br><!--bar--></span>{}</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><span><!--foo--><br><!--bar--></span>abc</p>" but got "<p>abc<span><!--foo--><!--bar--></span></p>"
[FAIL] [["inserthtml","<!--abc-->"]] "<p><span><!--foo--><br><!--bar--></span>{}</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><span><!--foo--><br><!--bar--></span><!--abc--></p>" but got "<p><!--abc--><span><!--foo--><!--bar--></span></p>"
[FAIL] [["inserthtml","<pre>abc</pre>"]] "<pre contenteditable=\\"false\\"><span contenteditable>[1234]</span></pre>" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<pre contenteditable=\\"false\\"><span contenteditable=\\"\\"><pre>abc</pre></span></pre>" not in array ["<pre contenteditable=\\"false\\"><span contenteditable=\\"\\">abc</span></pre>", "<pre contenteditable=\\"false\\"><span contenteditable=\\"\\">abc<br></span></pre>"]
[FAIL] [["inserthtml","<span></span>"]] "<div>a[]b</div>" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<div>ab</div>" not in array ["<div>a<span></span>b</div>", "<div>a<span></span>b<br></div>"]
[FAIL] [["inserthtml","<span class=\\"s1\\"></span>b<span class=\\"s2\\"></span>"]] "<div>a[]c</div>" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<div>ab</div><span class=\\"s1\\"></span><span class=\\"s2\\"></span><div>c</div>" not in array ["<div>a<span class=\\"s1\\"></span>b<span class=\\"s2\\"></span>c</div>", "<div>a<span class=\\"s1\\"></span>b<span class=\\"s2\\"></span>c<br></div>"]
[FAIL] [["inserthtml","<i>Z</i>"]] "<p><b><span contenteditable=false>abc</span>[Y]</b><i>def</i></p>" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<p><b><span contenteditable=\\"false\\">abc</span></b><i>def</i></p>" not in array ["<p><b><span contenteditable=\\"false\\">abc</span></b><i>Z</i><i>def</i></p>", "<p><b><span contenteditable=\\"false\\">abc</span></b><i>Zdef</i></p>"]
Harness: the test ran to completion.
